Team_TrueNorth/requirements.txt
django
django-allauth
requests
PyJWT
cryptography
requests
python-dotenv
google-generativeai
PyPDF2
python-docx
Pillow
